+++ web/components/rt/config_files/map.rt.js
@@ -1,0 +1,26 @@
+/*
+ * This is an iumio Framework component
+ *
+ * (c) RAFINA DANY <danyrafina@gmail.com>
+ *
+ * iumio Framework - iumio Components
+ *
+ * To get more information about licence, please check the licence file
+ */
+
+/* File generate by iumio JS Routing - Time : 2017-11-24 12:58:35. DO NOT OVERRIDE IT ! */
+
+var iumiortdata = {
+    "0": {
+        "HelloApp": []
+    },
+    "1": {
+        "DocApp": []
+    },
+    "2": {
+        "WhdeerzfrzzApp": []
+    },
+    "3": {
+        "demkdlmezfjelzfelkfnelkzjfkljeljkfApp": []
+    }
+};